refactor: numbers Input 중복 제거
